* 'feature/HabitTimer' of https://github.com/najak2007/HabitTimer:
  Pomodoro에서 시간 설정을 Picker로 처리되도록 수정

+++ HabitTimer/Model/PomodoroModel.swift
@@ -29,7 +29,7 @@ final class ToDoListData: Object, Comparable {
     @objc dynamic var isDone: Bool = false
     @objc dynamic var isRepeat: Bool = false
     @objc dynamic var placeName: String = ""
-    @objc dynamic var selectedMinute: Double = Config.POMODORO_WORK_TIME_MINUTE
+    @objc dynamic var selectedMinute: Int = Config.POMODORO_WORK_TIME_MINUTE
     @objc dynamic var remainingTime: Double = 0
     
     static func < (lhs: ToDoListData, rhs: ToDoListData) -> Bool {
